Rename test files to match src
