--- test/esp/ecl/Graph.ts
@@ -1,9 +1,0 @@
-import { expect } from "chai";
-
-import { ECLGraph } from "../../../src/esp/ecl/graph";
-
-describe("Graph", function () {
-    it("basic", function () {
-        expect(ECLGraph).to.be.a("function");
-    });
-});
