+++ test/esp/ecl/resource.ts
@@ -1,0 +1,9 @@
+import { expect } from "chai";
+
+import { Resource } from "../../../src/esp/ecl/resource";
+
+describe("Resource", function () {
+    it("basic", function () {
+        expect(Resource).to.be.a("function");
+    });
+});
